+++ Controller/src/se/jrat/controller/packets/outgoing/Packet12RemoteScreen.java
@@ -31,7 +31,7 @@ public class Packet12RemoteScreen extends AbstractOutgoingPacket {
 	}
 
 	@Override
-	public byte getPacketId() {
+	public short getPacketId() {
 		return 12;
 	}
